+++ CHANGELOG.md
@@ -10,6 +10,12 @@
     - Added `signature` view
     - Improved `struct` view to show an object or an array signature when expanded and `[i]` action button is hovered
     - Added `limit` option for `menu` view
+    - Improved `tree` view:
+        - Added `limitLines` option to limit lines count on root children render, limits to 25 lines by default; use `limitLines: false` to disable it
+        - Changed `limit` option to limit children count (this option served as `limitLines` before)
+        - Added `expanded` option to specify initial leaf expanded level
+        - Added `collapsable` option to hide a toggle element when a leaf is expaned
+        - Prevented large tree growing by keeping already visited data leafs collapsed
 
 ## 1.0.0-beta.8 (30-01-2019)